refactor: extract notification and admin shell wiring
